Miscellaneous code cleanup

+++ internal/time.go
@@ -181,7 +181,7 @@ func (v *timeVal) SecondsWithFraction() float64 {
 	// Timestamps that represent a date before the year 1678 or after 2262 can
 	// be represented as nanoseconds in an int64.
 	if 1678 < y && y < 2262 {
-		return float64(float64(t.UnixNano()) / 1000000000.0)
+		return float64(t.UnixNano()) / 1000000000.0
 	}
 	// Fall back to microsecond precision
 	us := t.Unix()*1000000 + int64(t.Nanosecond())/1000
